+++ common/platform_windows.go
@@ -11,10 +11,10 @@ import (
 
 const (
 	defaultRobocorpLocation = "%LOCALAPPDATA%\\robocorp"
-	defaultHoloLocation     = "c:\\ProgramData\\robocorp\\ht"
+	defaultHoloLocation     = "%ProgramData%\\robocorp\\ht"
 
 	defaultSema4Location     = "%LOCALAPPDATA%\\sema4ai"
-	defaultSema4HoloLocation = "c:\\ProgramData\\sema4ai\\ht"
+	defaultSema4HoloLocation = "%ProgramData%\\sema4ai\\ht"
 )
 
 var (